feat(paper): 实现试卷管理及AI组卷功能
- PaperController新增查询、手动创建及查询详情接口实现
- Paper实体中关联题目列表及分数排序实现
- PaperService接口新增自定义方法customPaperDetailById与customCreatePaper
- PaperServiceImpl实现试卷详情完整信息查询逻辑，包括题目和分数匹配
- 实现手动组卷方法，支持题目和分数映射及事务管理
- PaperQuestion中paperId类型由Integer改为Long
- PaperVo中题目ID映射类型由Integer改为Long
- 添加日志记录和异常处理，提升代码健壮性与可维护性

+++ exam-system-server-new/src/main/java/com/zmh/exam/entity/PaperQuestion.java
@@ -21,7 +21,7 @@ public class PaperQuestion extends BaseEntity {
     /**
      * 试卷ID
      */
-    private Integer paperId; // 试卷ID
+    private Long paperId; // 试卷ID
 
     /**
      * 题目ID
@@ -40,7 +40,7 @@ public class PaperQuestion extends BaseEntity {
      * @param questionId 题目ID
      * @param score 分数
      */
-    public PaperQuestion(Integer paperId, Long questionId, BigDecimal score) {
+    public PaperQuestion(Long paperId, Long questionId, BigDecimal score) {
         this.paperId = paperId;
         this.questionId = questionId;
         this.score = score;
