+++ exam-system-server-new/src/main/java/com/zmh/exam/vo/PaperVo.java
@@ -31,7 +31,7 @@ public class PaperVo implements Serializable {
 
     @Schema(description = "试卷题目配置，Key为题目ID，Value为该题分数", 
             example = "{\"1\": 5.0, \"2\": 10.0, \"3\": 15.0}")
-    private Map<Integer, BigDecimal> questions; // 题目ID及分值
+    private Map<Long, BigDecimal> questions; // 题目ID及分值
 
     private static final long serialVersionUID = 1L; // 序列化版本UID
 }
